% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\encoding{UTF-8}
\name{get.tbl}
\alias{get.tbl}
\title{Get table giving on the \code{url} and depth digits of structure}
\usage{
get.tbl(url, len)
}
\arguments{
\item{url}{character.}

\item{len}{integer. short coding digits.}
}
\value{
data.frame
}
\description{
Get table giving on the \code{url} and depth digits of structure
}
\examples{
url_tar  <- "http://www.stats.gov.cn/tjsj/tjbz/tjyqhdmhcxhfdm/2021/13/03/22/130322102.html"
l <- 12
out <- get.tbl(url = url_tar, len= l)
}
