% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_city.R
\docType{data}
\name{zone_city}
\alias{zone_city}
\title{City zoning  (level 2)}
\format{
data.frame
\describe{
\item{province}{ character }
\item{pid}{ character, province zoning id with 2 digits}
\item{url}{ character }
\item{cid}{ character, city zoning id with 4 digits  }
\item{cname}{ character, name of city}
}
}
\usage{
zone_city
}
\description{
City zoning  (level 2)
}
\keyword{datasets}
