% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_district.R
\docType{data}
\name{zone_district}
\alias{zone_district}
\title{District zoning (level 3)}
\format{
data.frame
\describe{
\item{province}{ character }
\item{pid}{ character, province zoning id with 2 digits}
\item{url}{ character }
\item{cid}{ character, city zoning id with 4 digits  }
\item{cname}{ character, name of city}
\item{did}{ character, district zoning id with 6 digits   }
\item{dname}{ character, name of district }
}
}
\usage{
zone_district
}
\description{
District zoning (level 3)
}
\keyword{datasets}
