% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_province.R
\docType{data}
\name{zone_province}
\alias{zone_province}
\title{Province zoning  (level 1)}
\format{
data.frame
\describe{
\item{province}{ character }
\item{pid}{ character, province zoning with 2 digits}
}
}
\usage{
zone_province
}
\description{
Province zoning  (level 1)
}
\keyword{datasets}
