% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoningr-package.R
\docType{package}
\name{zoningr-package}
\alias{zoningr}
\alias{zoningr-package}
\title{zoningr: Urban and rural districts zoning for statistic in China}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/huhuaping/zoningr/issues}
}

}
\author{
\strong{Maintainer}: Hu Huaping \email{huhuaping01@nwsuaf.edu.cn}

}
\keyword{internal}
